Issue-136: Fix implicit uint64 to int conversion

+++ streets_utils/streets_vehicle_list/include/vehicle_list.h
@@ -41,7 +41,7 @@ namespace streets_vehicles {
              * 
              * @param timeout time in milliseconds from current time after which vehicles will be removed from the vehicle list.
              */
-            void purge_old_vehicles(const int timeout);
+            void purge_old_vehicles(const uint64_t timeout);
             
             // Hide get_singleton method. Use static methods instead.
             using streets_singleton::get_singleton;
